fix(tests): move shared presentation test bases to :shared:test-utils, drop clientApp srcDirs graft breaking IDE resolution
