+++ apps/clientApp/src/androidUnitTest/kotlin/network/bisq/mobile/client/network/presentation/network/ClientNetworkOverviewContentTest.kt
@@ -8,8 +8,8 @@ import androidx.compose.ui.test.performScrollTo
 import network.bisq.mobile.client.common.test_utils.TestApplication
 import network.bisq.mobile.i18n.i18n
 import network.bisq.mobile.i18n.i18nPlural
-import network.bisq.mobile.presentation.common.test_utils.compose.BisqComposeUiTestBase
 import network.bisq.mobile.presentation.common.ui.components.network.NetworkHealthState
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Test
 import org.robolectric.annotation.Config
