+++ shared/presentation/src/androidUnitTest/kotlin/network/bisq/mobile/presentation/common/ui/components/molecules/dialog/ReconnectingOverlayUiTest.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.unit.dp
 import io.mockk.mockk
 import io.mockk.verify
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.common.test_utils.compose.BisqComposeUiTestBase
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Test
 
 /**
