+++ shared/presentation/src/androidUnitTest/kotlin/network/bisq/mobile/presentation/settings/payment_accounts/PaymentAccountsContentUiTest.kt
@@ -11,8 +11,8 @@ import io.mockk.verify
 import network.bisq.mobile.domain.model.account.fiat.UserDefinedFiatAccount
 import network.bisq.mobile.domain.model.account.fiat.UserDefinedFiatAccountPayload
 import network.bisq.mobile.i18n.i18n
-import network.bisq.mobile.presentation.common.test_utils.compose.BisqComposeUiTestBase
 import network.bisq.mobile.presentation.common.ui.utils.DataEntry
+import network.bisq.mobile.test.presentation.compose.BisqComposeUiTestBase
 import org.junit.Before
 import org.junit.Test
