+++ shared/test-utils/src/androidMain/kotlin/network/bisq/mobile/test/presentation/compose/BisqComposeUiTestBase.kt
@@ -1,10 +1,10 @@
-package network.bisq.mobile.presentation.common.test_utils.compose
+package network.bisq.mobile.test.presentation.compose
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import network.bisq.mobile.i18n.I18nSupport
-import network.bisq.mobile.presentation.common.test_utils.compose.BisqComposeTestSupport.setBisqTestContent
+import network.bisq.mobile.test.presentation.compose.BisqComposeTestSupport.setBisqTestContent
 import org.junit.Before
 import org.junit.Rule
 import org.junit.runner.RunWith
