+++ shared/test-utils/src/androidMain/kotlin/network/bisq/mobile/test/presentation/coroutines/PlatformPresentationKoinTestBase.kt
@@ -1,4 +1,4 @@
-package network.bisq.mobile.presentation.common.test_utils.coroutines
+package network.bisq.mobile.test.presentation.coroutines
 
 abstract class PlatformPresentationKoinTestBase : PresentationKoinTestBase() {
     override fun setUpPlatformMocks() {
